Refactor crud for User and Application with related imports.
